@@ -1,102 +1,64 @@
 {
-  "name": "@prisma-cms/boilerplate-modx",
+  "name": "@modxclub.ru/boilerplate",
   "version": "1.0.0",
-  "description": "@prisma-cms/boilerplate for MODX Revolution",
+  "description": "",
   "private": false,
   "scripts": {
     "start": "react-scripts start",
     "build": "PUBLIC_URL=/ react-scripts build",
     "build-less": "node-less-chokidar src/styles/less",
-    "start-ssr": "NODE_ENV=production node src/ssr",
+    "start-ssr": "NODE_ENV=production node --experimental-modules src/ssr",
     "start-ssr-dev": "NODE_ENV=production nodemon src/ssr",
-    "start-server": "Sendmail=true action=start-server node src/server/",
-    "start-server-dev": "Sendmail=true SendmailTest=true action=start-server nodemon src/server/index.js",
-    "deploy": "NODE_ENV=test node src/scripts/deploy/with-prisma",
-    "build-api": "node src/scripts/deploy/api-only",
+    "start-server": "Sendmail=true action=start-server node --experimental-modules src/server/",
+    "start-server-dev": "Sendmail=true SendmailTest=true action=start-server nodemon --experimental-modules src/server/",
+    "deploy": "NODE_ENV=test node --experimental-modules src/server/scripts/deploy/with-prisma",
+    "build-api": "node --experimental-modules src/server/scripts/deploy/api-only",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/register": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "bcryptjs": "^2.4.3",
-    "graphql": "14.0.2",
-    "graphql-cli": "^2.15.13",
-    "graphql-cli-generate-fragments": "^1.4.1",
-    "graphql-tag": "2.10.0",
-    "graphql-yoga": "^1.16.1",
-    "html-to-json": "^0.6.0",
-    "i18n": "^0.8.3",
-    "i18next-browser-languagedetector": "^2.2.3",
-    "knex": "^0.15.2",
-    "merge-graphql-schemas": "^1.5.7",
-    "mysql": "^2.16.0",
-    "php-session-unserialize": "^0.2.1",
-    "prettier-eslint-cli": "4.7.1",
-    "prisma-binding": "^2.1.5",
-    "prisma-cli-engine": "^1.19.0",
-    "promises-all": "^1.0.0",
-    "roddeh-i18n": "^1.1.1"
-  },
   "dependencies": {
-    "@babel/polyfill": "^7.0.0",
-    "@babel/runtime": "^7.1.2",
     "@modxclub/react-editor": "^1.0.3",
-    "@prisma-cms/component": "^1.0.7",
+    "@prisma-cms/component": "^1.0.9",
     "@prisma-cms/e-commerce": "^1.0.1",
-    "@prisma-cms/front": "^1.9.6-beta",
-    "@prisma-cms/server": "^1.6.7-beta",
-    "@prisma-cms/ui": "^1.2.0",
+    "@prisma-cms/front": "^1.9.8-beta",
+    "@prisma-cms/log-module": "^1.1.0-beta",
+    "@prisma-cms/mail-module": "^1.1.0-beta",
+    "@prisma-cms/server": "^2.1.7-beta",
+    "@prisma-cms/ui": "^1.2.2",
+    "@prisma-cms/upload-module": "^1.3.5-beta",
+    "@prisma-cms/user-module": "^1.1.0-beta",
     "@types/react": "^16.1.0",
-    "apollo-boost": "^0.1.17",
-    "apollo-cache-inmemory": "^1.3.6",
-    "apollo-client": "^2.4.3",
-    "apollo-error": "^1.0.2",
-    "apollo-fetch-upload": "^1.2.0",
-    "apollo-link": "^1.2.3",
-    "apollo-link-batch-http": "^1.0.4",
-    "apollo-link-error": "^1.1.1",
+    "apollo-cache-inmemory": "^1.3.7",
+    "apollo-client": "^2.4.4",
     "apollo-link-http": "1.5.5",
-    "apollo-link-ws": "^1.0.9",
-    "apollo-upload-client": "^9.0.0",
-    "apollo-utilities": "^1.0.23",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "debug": "^4.1.0",
-    "dom-scroll-into-view": "^1.2.1",
-    "draft-js": "^0.10.5",
-    "graphql": "14.0.2",
-    "graphql-cli": "^2.15.13",
-    "graphql-tag": "2.6.1",
+    "cheerio": "^1.0.0-rc.2",
     "graphql-type-json": "^0.2.1",
-    "jss-theme-reactor": "^0.11.1",
-    "load-script": "^1.0.0",
     "material-ui": "1.0.0-beta.43",
     "material-ui-icons": "^1.0.0-beta.17",
     "moment": "^2.20.1",
     "moment-timezone": "^0.5.14",
     "node-less-chokidar": "^0.3.1",
     "node-sass-chokidar": "^1.1.0",
+    "php-session-unserialize": "^0.2.1",
     "prisma-binding": "^2.1.6",
-    "react": "16.5.2",
+    "react": "^16.6.0",
     "react-apollo": "^2.2.3",
-    "react-cms-graphql-utils": "^1.1.0",
-    "react-dom": "^16.5.2",
-    "react-router-dom": "^4.3.1",
+    "react-dom": "^16.6.0",
     "react-scripts": "^2.0.2",
-    "shortid": "^2.2.8",
-    "socket.io": "^2.0.4",
-    "socket.io-parser": "^3.1.2",
-    "subscriptions-transport-ws": "^0.9.5",
     "urijs": "^1.19.1",
     "xml-writer": "^1.7.0"
   },
-  "resolutions": {
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
+  "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "graphql": "14.0.2",
-    "prisma-binding": "^2.1.6"
+    "graphql-tag": "2.10.0",
+    "merge-graphql-schemas": "^1.5.7"
+  },
+  "resolutions": {
+    "graphql": "14.0.2"
   },
   "browserslist": [
     ">0.2%",
