@@ -1,6 +1,7 @@
 
-const sharp = require('sharp');
-var fs = require('fs');
+import sharp from "sharp";
+
+import fs from "fs";
 
 
 const resizeMax = async function (img, width, height) {
@@ -37,10 +38,6 @@ const ImageThumbMiddleware = async (req, res, next) => {
   } = req.params;
 
   let path = `/uploads/${src}`;
-
-  // const abthPath = __dirname + path;
-
-  // console.log("ImageThumbMiddleware", path);
 
   const abthPath = process.cwd() + path;
 
@@ -145,4 +142,4 @@ const ImageThumbMiddleware = async (req, res, next) => {
 
 }
 
-module.exports = ImageThumbMiddleware;
+export default ImageThumbMiddleware;
